add templates for users/items move methods for settings to a separate view file hide login button if user authenticated
